Fix audit findings: integer overflow, ops array size, VLA guard, test leak
C1 (critical): repack_q4_0_for_gpu n_blocks used int*int which can
overflow for large matrices. Changed to size_t arithmetic.

H2 (high): GPU ops array max_ops=23*n_layers could underflow for models
with all 3 Q/K/V biases (up to 25 ops/layer). Bumped to 26*n_layers.

M5 (medium): forward_gpu had VLA float emb[dim] without the
BN_MAX_VLA_ELEMS guard that forward_layers has. Added.

M10 (medium): test_coherence leaked mapped file on early error returns.
Added cleanup on GGUF parse and model load failures.

+++ src/gpu_wgpu.c
@@ -485,7 +485,7 @@ static void *repack_q4_0_for_gpu(BnWgpuCtx *ctx, const void *data, size_t size,
                                    const float *bias, int bias_len)
 {
     (void)size;
-    int n_blocks = rows * (cols / 32);
+    int n_blocks = (int)((size_t)rows * ((unsigned)cols / 32));
     size_t base_size = (size_t)n_blocks * 4  /* f32 scales */
                      + (size_t)n_blocks * 16; /* nibble data (4 u32s per block) */
     size_t bias_bytes = (bias && bias_len > 0) ? (size_t)bias_len * sizeof(float) : 0;
